improved loging

+++ k8s_HA/provision.sh
@@ -1,44 +1,73 @@
-echo "Updating VM..."
-apt-get update -y && apt-get upgrade -y
-echo "Done !"
+log() {
+  echo -e "[$(date +'%Y-%m-%d %H:%M:%S')] [INFO] $1"
+}
 
-echo "Installing default tools"
-apt-get install curl wget vim net-tools -y
-echo "Done !"
+ok() {
+  echo -e "[$(date +'%Y-%m-%d %H:%M:%S')] [OK]   $1"
+}
 
-if ! hostname | grep "loadbalancer"; then
-  echo "Setting up kubectl"
-  curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl"
-  install -o root -g root -m 0755 kubectl /usr/local/bin/kubectl
-  rm kubectl
-  echo "Done !"
+error() {
+  echo -e "[$(date +'%Y-%m-%d %H:%M:%S')] [ERROR] $1" >&2
+}
 
-  echo "Adding alias 'k' for 'kubectl'"
+log "Updating and upgrading system packages..."
+if apt-get update -y && apt-get upgrade -y; then
+  ok "System updated successfully."
+else
+  error "System update failed."
+fi
+
+log "Installing default tools: curl, wget, vim, net-tools..."
+if apt-get install curl wget vim net-tools -y; then
+  ok "Default tools installed."
+else
+  error "Tool installation failed."
+fi
+
+if ! hostname | grep -q "loadbalancer"; then
+  log "Setting up kubectl..."
+
+  if curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl" &&
+     install -o root -g root -m 0755 kubectl /usr/local/bin/kubectl &&
+     rm kubectl; then
+    ok "kubectl installed."
+  else
+    error "kubectl installation failed."
+  fi
+
+  log "Adding alias 'k' for 'kubectl' in .bashrc"
   echo "alias k='kubectl'" >> /home/vagrant/.bashrc
   chown vagrant:vagrant /home/vagrant/.bashrc
 
-  echo "Adding auto-completion for 'kubectl' and 'k' commands"
+  log "Enabling autocompletion for 'kubectl' and 'k'"
   echo "source <(kubectl completion bash)" >> /home/vagrant/.bashrc
   echo "complete -F __start_kubectl k" >> /home/vagrant/.bashrc
-  echo "Done !"
+  ok "Alias and autocompletion configured."
 fi
 
 if hostname | grep -q "master-1"; then
-  su - vagrant -c 'ssh-keygen -t rsa -b 2048 -f ~/.ssh/id_rsa -q -N ""'
+  log "Generating SSH key for vagrant user on master-1..."
+  if su - vagrant -c 'ssh-keygen -t rsa -b 2048 -f ~/.ssh/id_rsa -q -N ""'; then
+    ok "SSH key generated."
 
-  cp /home/vagrant/.ssh/id_rsa.pub /vagrant/master1.pub
-  chown vagrant:vagrant /vagrant/master1.pub
+    cp /home/vagrant/.ssh/id_rsa.pub /vagrant/master1.pub
+    chown vagrant:vagrant /vagrant/master1.pub
+    ok "Public key copied to /vagrant/master1.pub"
+  else
+    error "SSH key generation failed."
+  fi
 fi
 
 if [ -f /vagrant/master1.pub ]; then
+  log "Adding master-1 public key to authorized_keys..."
   mkdir -p /home/vagrant/.ssh
   cat /vagrant/master1.pub >> /home/vagrant/.ssh/authorized_keys
   chown -R vagrant:vagrant /home/vagrant/.ssh
   chmod 600 /home/vagrant/.ssh/authorized_keys
+  ok "Public key added to authorized_keys."
 fi
 
-echo "[INFO] Updating /etc/hosts for cluster node resolution"
-
+log "Updating /etc/hosts for cluster node resolution..."
 cat <<EOF >> /etc/hosts
 192.168.56.11 m1
 192.168.56.12 m2
@@ -48,3 +77,4 @@ cat <<EOF >> /etc/hosts
 192.168.56.23 w3
 192.168.56.30 lb
 EOF
+ok "/etc/hosts updated."
